Separated render initialization code from main.js
